Split leveled logging out of Utils into Log.h/Log.cpp
