+++ DeadSpaceFixes/Utils.cpp
@@ -1,52 +1,16 @@
 #define WIN32_LEAN_AND_MEAN
 
 #include <Windows.h>
-#include <iostream>
 #include <vector>
 #include <cstdint>
-#include <cstdarg>
+#include <cstring>
+#include <cstdlib>
 #include <emmintrin.h>
 #include <intrin.h>
 
 #include "Utils.h"
 
 namespace Utils {
-    void InitialiseConsole()
-    {
-        AllocConsole();
-        FILE* fDummy;
-        freopen_s(&fDummy, "CONOUT$", "w", stdout);
-        freopen_s(&fDummy, "CONOUT$", "w", stderr);
-    }
-
-    namespace {
-        // Only messages at or above this level are printed. Defaults to Info so
-        // the verbose "Found <addr>" DEBUG lines stay out of the way.
-        LogLevel g_LogLevel = LogLevel::Info;
-    }
-
-    void SetLogLevel(LogLevel level)
-    {
-        g_LogLevel = level;
-    }
-
-    void Log(LogLevel level, const char* /*tag*/, const char* fmt, ...)
-    {
-        if (level < g_LogLevel)
-            return;
-
-        static const char* kLevelNames[] = { "DBG", "INF", "WRN", "ERR" };
-        const char* levelNames = (level >= LogLevel::Debug && level <= LogLevel::Error)
-            ? kLevelNames[static_cast<int>(level)] : "???";
-
-        printf("[DeadSpaceFixes] [%s] ", levelNames);
-
-        va_list args;
-        va_start(args, fmt);
-        vprintf(fmt, args);
-        va_end(args);
-        printf("\n");
-    }
 
     //Converts a signature string like "8B ? 50 F3 0F" into a byte array where -1 means
     //"any byte" (the ? wildcard). Extracted into its own helper so the whole-image and
